composite PK applied in CommentVote, ArticleVote
