+++ src/main/kotlin/org/oooc/kry/board/article/comment/domain/entity/CommentVote.kt
@@ -1,0 +1,21 @@
+package org.oooc.kry.board.domain
+
+import javax.persistence.*
+
+@Entity
+@IdClass(CommentVoteId::class)
+data class CommentVote (
+
+    @Id
+    @ManyToOne
+    var comment: Comment,
+
+//    TODO: Mapping to User
+//    @ManyToOne
+//    @Id
+//    var user: User = User(),
+
+    @Id
+    var updown: Long
+
+)
