+++ src/main/kotlin/org/oooc/kry/board/article/comment/domain/entity/CommentVoteId.kt
@@ -1,0 +1,9 @@
+package org.oooc.kry.board.domain
+
+import java.io.Serializable
+
+data class CommentVoteId (
+    var comment: String,
+    var user: String,
+    var updown: String
+): Serializable
