+++ src/main/kotlin/org/oooc/kry/board/article/domain/entity/ArticleVoteId.kt
@@ -1,0 +1,9 @@
+package org.oooc.kry.board.domain
+
+import java.io.Serializable
+
+data class ArticleVoteId (
+    var article: String,
+    // var user: String,
+    var updown: String
+): Serializable
