+++ src/main/kotlin/org/oooc/kry/board/domain/dto/BoardGetRequestDTO.kt
@@ -1,0 +1,5 @@
+package org.oooc.kry.board.dto
+
+data class BoardGetRequestDTO (
+    val boardName: String
+)
