--- src/main/kotlin/org/oooc/kry/board/dto/BoardGetRequestDTO.kt
@@ -1,5 +1,0 @@
-package org.oooc.kry.board.dto
-
-data class BoardGetRequestDTO (
-    val boardName: String
-)
